Fix objects jitter when applying transformations.
When we apply transformations to the objects using mouse drag, the
affected object seem to jitter because transformation is immediately
applied on `observer` call. Which not necessary happens on the right
time in call schedule.

By extracting transform application to the separate function that runs
on update jittering issue is fixed.